+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/VariantTypesClasses/DeleteVariantTypesClass/DeleteVariantTypesClassCommandValidator.cs
@@ -1,0 +1,24 @@
+using System;
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "1.0")]
+
+namespace CleanArchitecture.TestApplication.Application.VariantTypesClasses.DeleteVariantTypesClass
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class DeleteVariantTypesClassCommandValidator : AbstractValidator<DeleteVariantTypesClassCommand>
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore, Signature = Mode.Merge)]
+        public DeleteVariantTypesClassCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [IntentManaged(Mode.Fully)]
+        private void ConfigureValidationRules()
+        {
+        }
+    }
+}