+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Infrastructure/DependencyInjection.cs
@@ -28,6 +28,7 @@ namespace CleanArchitecture.TestApplication.Infrastructure
             services.AddTransient<IAggregateRootLongRepository, AggregateRootLongRepository>();
             services.AddTransient<IAggregateSingleCRepository, AggregateSingleCRepository>();
             services.AddTransient<IImplicitKeyAggrRootRepository, ImplicitKeyAggrRootRepository>();
+            services.AddTransient<IVariantTypesClassRepository, VariantTypesClassRepository>();
             services.AddScoped<IDomainEventService, DomainEventService>();
             return services;
         }